Reorganize server.ts into focused modules
server.ts was carrying several unrelated concerns. Split them out so the
file is just MCP server/tool wiring (497 -> 123 lines):

- src/tools/: docs-search.ts (moved) + non-codemode.ts (new, holds the
  ~165-line registerNonCodemodeTools REST passthrough).
- src/metrics.ts: gains attachMetrics + userIdFromProps and now owns
  SERVER_INFO (its natural home as the metrics client identity).
- src/descriptions.ts (new): all tool description / prompt text and the
  CLOUDFLARE_TYPES declarations, exposed as searchToolDescription,
  executeToolDescription, and accountIdParamDescription builders.
- src/utils/errors.ts (new): formatError.

No behavior change; tests updated for the moved docs-search path.
