+++ src/tools/docs-search.ts
