+++ src/utils/errors.ts
@@ -1,0 +1,6 @@
+/**
+ * Normalize any thrown value to a human-readable message.
+ */
+export function formatError(error: unknown): string {
+  return error instanceof Error ? error.message : String(error)
+}
